fix: 🐛 fix routing and rerenders, memoize routes and session
