+++ src/components/app/ts/app.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { useContext, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { Router } from 'react-router-dom';
+import { Stomp } from '@stomp/stompjs';
+import * as SockJS from 'sockjs-client';
 import { createBrowserHistory } from 'history';
-import { AppRouter } from './AppRouter';
+import { Routing } from './Routing';
 import { config } from '../../../resources/ts/config';
 import {
 	setTokenInCookie,
@@ -14,12 +16,12 @@ import {
 	UserDataInterface,
 	AuthDataContext,
 	AuthDataInterface,
-	hasUserAuthority,
-	AUTHORITIES
+	UnreadSessionsStatusContext
 } from '../../../globalState';
 import { ContextProvider } from '../../../globalState/state';
 import { getUserData } from '../../apiWrapper/ts/';
 import { Loading } from './Loading';
+import { logout } from '../../logout/ts/logout';
 
 export const history = createBrowserHistory();
 
@@ -35,28 +37,25 @@ export const AppContainer = (props) => {
 	);
 };
 
-export const App = (props) => {
+const socket = new SockJS(config.endpoints.liveservice);
+const stompClient = Stomp.over(socket);
+// implement for release to deactivate stomp logging
+// stompClient.debug = null;
+
+const STOMP_EVENT_TYPES = {
+	DIRECT_MESSAGE: 'directMessage'
+};
+
+export const App = () => {
 	const { setAuthData } = useContext(AuthDataContext);
 	const [authDataRequested, setAuthDataRequested] = useState(false);
-	const { userData, setUserData } = useContext(UserDataContext);
+	const { setUserData } = useContext(UserDataContext);
 	const [appReady, setAppReady] = useState(false);
 	const [userDataRequested, setUserDataRequested] = useState(false);
-
-	const renderCondition = () => {
-		if (
-			hasUserAuthority(AUTHORITIES.VIEW_ALL_PEER_SESSIONS, userData) ||
-			hasUserAuthority(AUTHORITIES.CONSULTANT_DEFAULT, userData) ||
-			hasUserAuthority(AUTHORITIES.USER_DEFAULT, userData)
-		) {
-			return <AppRouter />;
-		}
-		window.location.href = config.endpoints.logoutRedirect;
-		return null;
-	};
-
-	if (appReady) {
-		return <Router history={history}>{renderCondition()}</Router>;
-	}
+	const { unreadSessionsStatus, setUnreadSessionsStatus } = useContext(
+		UnreadSessionsStatusContext
+	);
+	const [newStompDirectMessage, setNewStompDirectMessage] = useState(false);
 
 	if (!authDataRequested) {
 		setAuthDataRequested(true);
@@ -82,8 +81,63 @@ export const App = (props) => {
 				setAppReady(true);
 			})
 			.catch((error) => {
+				window.location.href = config.endpoints.logoutRedirect;
 				console.log(error);
 			});
+	}
+
+	useEffect(() => {
+		initLiveServiceSocket();
+	}, [appReady]);
+
+	useEffect(() => {
+		if (newStompDirectMessage) {
+			setUnreadSessionsStatus({
+				mySessions: unreadSessionsStatus.mySessions + 1,
+				newDirectMessage: true,
+				resetedAnimations: unreadSessionsStatus.mySessions === 0
+			});
+			setNewStompDirectMessage(false);
+		}
+	}, [newStompDirectMessage]);
+
+	const initLiveServiceSocket = () => {
+		stompClient.connect(
+			{
+				accessToken: getTokenFromCookie('keycloak')
+			},
+			(frame) => {
+				console.log('Connected: ' + frame);
+				stompClient.subscribe('/user/events', function (message) {
+					const stompMessageBody = JSON.parse(message.body);
+					const stompEventType = stompMessageBody['eventType'];
+					if (stompEventType === STOMP_EVENT_TYPES.DIRECT_MESSAGE) {
+						setNewStompDirectMessage(true);
+					}
+				});
+			}
+		);
+		stompClient.onWebSocketClose = (message) => {
+			console.log('Closed', message);
+		};
+		stompClient.onWebSocketError = (error) => {
+			console.log('Error', error);
+		};
+	};
+
+	const handleLogout = () => {
+		if (stompClient) {
+			stompClient.disconnect();
+		}
+		logout();
+	};
+
+	if (appReady) {
+		return (
+			<Router history={history}>
+				<Routing logout={handleLogout} />
+			</Router>
+		);
 	}
 
 	return <Loading />;
