+++ src/components/messageSubmitInterface/ts/messageSubmitInterfaceComponent.tsx
@@ -118,7 +118,7 @@ export interface MessageSubmitInterfaceComponentProps {
 	placeholder: string;
 	showMonitoringButton?: Function;
 	type: string;
-	typingUsers?: [];
+	typingUsers?: any[];
 }
 
 export const MessageSubmitInterfaceComponent = (
